sql commit

+++ web/WEB-INF/jsp/index.jsp
@@ -124,7 +124,7 @@
                                     <a href="<%= request.getContextPath()%>/editStudent/${row.id}" class="btn btn-warning"><i class="fa fa-edit"></i> Edit</a>
                                 </td>
                                 <td>
-                                    <a href="<%= request.getContextPath()%>/deleteStudent/${row.id}" class="btn btn-danger"><i class="fa fa-trash"></i> Delete</a>
+                                    <a onclick="return confirm('Are you want to delete this item?')" href="<%= request.getContextPath()%>/deleteStudent/${row.id}" class="btn btn-danger"><i class="fa fa-trash"></i> Delete</a>
                                 </td>
                             </tr>
                         </c:forEach>
